// packages/tds-widget/src/poi-detail/types.ts
export type ImageCategoryOrder =
  | 'image' // 대표 이미지
  | 'recommendation'
  | 'menuItem' // 대표 메뉴 이미지
  | 'menuBoard' // 메뉴판 이미지
  | 'featuredContent'
  | 'images' // 어드민에 등록된 이미지들 가운데 대표 이미지를 제외한 나머지 이미지